feat(metrics): add students distributions in summary report
